Use `spark_df_to_pandas` in preview also. Refactoring.
